+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/snapshot/ProjectGitSnapshotManager.kt
@@ -1,35 +1,27 @@
 package org.plan.research.minimization.plugin.snapshot
 
-import arrow.core.raise.either
-import arrow.core.raise.recover
-import com.intellij.openapi.application.writeAction
-import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.project.ex.ProjectManagerEx
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.NonCancellable
-import kotlinx.coroutines.withContext
-import mu.KotlinLogging
 import org.plan.research.minimization.plugin.errors.SnapshotError
 import org.plan.research.minimization.plugin.errors.SnapshotError.*
 import org.plan.research.minimization.plugin.logging.statLogger
-import org.plan.research.minimization.plugin.model.HeavyIJDDContext
 import org.plan.research.minimization.plugin.model.IJDDContext
 import org.plan.research.minimization.plugin.model.snapshot.SnapshotManager
 import org.plan.research.minimization.plugin.model.snapshot.TransactionBody
 import org.plan.research.minimization.plugin.model.snapshot.TransactionResult
-import org.plan.research.minimization.plugin.services.ProjectCloningGitService
-import org.plan.research.minimization.plugin.services.ProjectCloningService
-import org.eclipse.jgit.api.Git
-import org.eclipse.jgit.api.ResetCommand
+import org.plan.research.minimization.plugin.services.GitWrapperService
 
-class ProjectCloningGitSnapshotManager(rootProject: Project) : SnapshotManager {
-    private val projectCloning = rootProject.service<ProjectCloningGitService>()
+import arrow.core.raise.either
+import arrow.core.raise.recover
+import com.intellij.openapi.components.service
+import com.intellij.openapi.project.Project
+import mu.KotlinLogging
+
+class ProjectGitSnapshotManager(rootProject: Project) : SnapshotManager {
+    private val gitWrapperService = rootProject.service<GitWrapperService>()
     private val generalLogger = KotlinLogging.logger {}
 
     override suspend fun <T, C : IJDDContext> transaction(
         context: C,
-        action: suspend TransactionBody<T>.(newContext: C) -> C
+        action: suspend TransactionBody<T>.(newContext: C) -> C,
     ): TransactionResult<T, C> = either {
         statLogger.info { "Snapshot manager start's transaction" }
         generalLogger.info { "Snapshot manager start's transaction" }
@@ -45,27 +37,14 @@ class ProjectCloningGitSnapshotManager(rootProject: Project) : SnapshotManager {
                 catch = { raise(TransactionFailed(it)) },
             )
         } catch (e: Throwable) {
-            println("reset project")
-            resetProject(context)
+            gitWrapperService.resetChanges(context)
             throw e
         }
     }.onRight {
         generalLogger.info { "Transaction completed successfully" }
         statLogger.info { "Transaction result: success" }
-        println("commit project")
-        projectCloning.commitChanges(context)
+        gitWrapperService.commitChanges(context)
     }.onLeft { it.log() }
-
-    private suspend fun resetProject(context: IJDDContext) {
-        withContext(Dispatchers.IO) { // git reset --hard
-            Git.open(context.projectDir.toNioPath().toFile()).apply {
-                println("HEAD reset ${this.repository.resolve("HEAD")?.name}")
-                reset().setMode(ResetCommand.ResetType.HARD).call()
-                clean().setCleanDirectories(true).call()
-                close()
-            }
-        }
-    }
 
     private fun <T> SnapshotError<T>.log() = when (this) {
         is Aborted<*> -> {
