+++ project-minimization-plugin/src/test/kotlin/ProjectCloningGitSnapshotTest.kt
@@ -3,27 +3,21 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.vfs.*
 import com.intellij.testFramework.utils.vfs.deleteRecursively
-import junit.framework.TestCase
 import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withContext
 import org.eclipse.jgit.api.Git
 import org.eclipse.jgit.api.ResetCommand
 import org.eclipse.jgit.api.errors.NoHeadException
 import org.eclipse.jgit.treewalk.TreeWalk
-import org.junit.jupiter.api.assertDoesNotThrow
 import org.plan.research.minimization.plugin.errors.SnapshotError
 import org.plan.research.minimization.plugin.getAllNestedElements
 import org.plan.research.minimization.plugin.model.HeavyIJDDContext
 import org.plan.research.minimization.plugin.model.IJDDContext
 import org.plan.research.minimization.plugin.model.LightIJDDContext
-import org.plan.research.minimization.plugin.services.ProjectCloningGitService
-import org.plan.research.minimization.plugin.services.ProjectCloningService
-import org.plan.research.minimization.plugin.snapshot.ProjectCloningGitSnapshotManager
+import org.plan.research.minimization.plugin.services.GitWrapperService
+import org.plan.research.minimization.plugin.snapshot.ProjectGitSnapshotManager
 import java.io.File
 import kotlin.io.path.Path
 import kotlin.io.path.relativeTo
-import kotlin.reflect.full.declaredFunctions
-import kotlin.reflect.jvm.isAccessible
 
 abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBaseTest(), TestWithContext<C> {
     fun testOneFileProjectPartialCloning() {
@@ -75,10 +69,11 @@ abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBa
     ) {
         val project = myFixture.project
         val projectDir = project.guessProjectDir()!!
-        val snapshotManager = ProjectCloningGitSnapshotManager(project)
-        val git = gitInitOrOpen()
-        runBlocking { ProjectCloningGitService().commitChanges(createContext(project)) }
-        val originalCommitList = getCommitList(git)
+        val snapshotManager = ProjectGitSnapshotManager(project)
+        val gitWrapperService = project.service<GitWrapperService>()
+        val git = gitWrapperService.gitInitOrOpen(projectDir)
+        runBlocking { gitWrapperService.commitChanges(createContext(project)) }
+        val originalCommitList = gitWrapperService.getCommitList(git)
         val originalFiles =
             selectedFiles.getAllFiles(projectDir) + projectDir.getPathContentPair(projectDir.toNioPath()) +
                     projectDir.findChild(".git")!!.getAllFiles(projectDir.toNioPath())
@@ -102,7 +97,7 @@ abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBa
         }.getOrNull()
         assertNotNull(clonedProject)
         val clonedFiles = clonedProject!!.projectDir.getAllFiles(clonedProject.projectDir.toNioPath())
-        val clonedCommitList = getCommitList(git)
+        val clonedCommitList = gitWrapperService.getCommitList(git)
 
         assertEquals(originalFiles.filter { !it.path.startsWith(".git") && !it.path.toString().contains("/.git/")}.toSet(),
             clonedFiles.filter { !it.path.startsWith(".git") && !it.path.toString().contains("/.git/")}.toSet())
@@ -116,11 +111,13 @@ abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBa
     fun testAbortedTransaction() {
         myFixture.copyDirectoryToProject("flatProject", "")
         val project = myFixture.project
-        val git = gitInitOrOpen()
-        val snapshotManager = ProjectCloningGitSnapshotManager(project)
+        val projectDir = project.guessProjectDir()!!
+        val gitWrapperService = project.service<GitWrapperService>()
+        val git = gitWrapperService.gitInitOrOpen(projectDir)
+        val snapshotManager = ProjectGitSnapshotManager(project)
         val context = createContext(project)
-        runBlocking { ProjectCloningGitService().commitChanges(createContext(project)) }
-        val originalCommitList = getCommitList(git)
+        runBlocking { gitWrapperService.commitChanges(createContext(project)) }
+        val originalCommitList = gitWrapperService.getCommitList(git)
         val result = runBlocking {
             snapshotManager.transaction(context) { newContext ->
                 writeAction {
@@ -132,7 +129,7 @@ abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBa
         assertEquals(SnapshotError.Aborted("Abort"), result)
         project.guessProjectDir()!!.refresh(false, true)
         assertNotNull(project.guessProjectDir()!!.findChild(".config"))
-        val clonedCommitList = getCommitList(git)
+        val clonedCommitList = gitWrapperService.getCommitList(git)
         assertEquals(originalCommitList.size, clonedCommitList.size)
         assert(project.isOpen)
     }
@@ -140,12 +137,14 @@ abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBa
     fun testExceptionInTransaction() {
         myFixture.copyDirectoryToProject("flatProject", "")
         val project = myFixture.project
-        val git = gitInitOrOpen()
+        val projectDir = project.guessProjectDir()!!
+        val gitWrapperService = project.service<GitWrapperService>()
+        val git = gitWrapperService.gitInitOrOpen(projectDir)
         val context = createContext(project)
-        runBlocking { ProjectCloningGitService().commitChanges(createContext(project)) }
-        val originalCommitList = getCommitList(git)
+        runBlocking { gitWrapperService.commitChanges(createContext(project)) }
+        val originalCommitList = gitWrapperService.getCommitList(git)
 
-        val snapshotManager = ProjectCloningGitSnapshotManager(project)
+        val snapshotManager = ProjectGitSnapshotManager(project)
         val result = runBlocking {
             snapshotManager.transaction<String, _>(context) { newContext ->
                 writeAction {
@@ -158,66 +157,9 @@ abstract class ProjectCloningGitSnapshotTest<C : IJDDContext> : ProjectCloningBa
         assertEquals("Abort", (result as? SnapshotError.TransactionFailed)?.error?.message)
         project.guessProjectDir()!!.refresh(false, true)
         assertNotNull(project.guessProjectDir()!!.findChild(".config"))
-        val clonedCommitList = getCommitList(git)
+        val clonedCommitList = gitWrapperService.getCommitList(git)
         assertEquals(originalCommitList.size, clonedCommitList.size)
         assert(project.isOpen)
-    }
-
-    private fun gitInitOrOpen(): Git {
-        val projectDir: File = myFixture.project.guessProjectDir()!!.toNioPath().toFile()
-        if (projectDir.resolve(".git").exists()) {
-            println("open project")
-            return Git.open(projectDir)
-        }
-        println("init project")
-        return Git.init()
-            .setDirectory(myFixture.project.guessProjectDir()!!.toNioPath().toFile())
-            .call().apply {
-                commit().setMessage("init commit").call()
-            }
-    }
-
-    private fun commitListEmpty(git: Git): Boolean {
-        return try {
-            git.log().call() // This will throw NoHeadException if there are no commits
-            false
-        } catch (e: NoHeadException) {
-            true
-        } finally {
-            git.close()
-        }
-    }
-
-    /* returns all commits in the chronically reversed order */
-    /* getCommitList(git)[0] == HEAD */
-    private fun getCommitList(git: Git): List<String> {
-        myFixture.project.guessProjectDir()!!.refresh(false, true)
-        if (commitListEmpty(git)) {
-            return listOf()
-        }
-        return git.log().all().call().map { it.name }
-    }
-
-    private fun getTrackedFiles(): List<String> {
-        val repoDir = File(myFixture.project.guessProjectDir()!!.path)
-        val git = Git.open(repoDir)
-        val repository = git.repository
-        val trackedFiles = mutableListOf<String>()
-
-        // Resolve HEAD to the latest commit
-        val headCommit = repository.resolve("HEAD^{tree}")
-
-        // Traverse the tree of the latest commit
-        TreeWalk(repository).use { treeWalk ->
-            treeWalk.addTree(headCommit)
-            treeWalk.isRecursive = true // List all files recursively
-            while (treeWalk.next()) {
-                trackedFiles.add(treeWalk.pathString)
-            }
-        }
-
-        git.close()
-        return trackedFiles
     }
 }
